Complete SOLID domain ownership remediation
Squash merge of the validated SOLID remediation and CI-source isolation refinements.

+++ src/domains/kv/actor/mutations.rs
@@ -26,9 +26,7 @@ impl KvActor {
                 found: false,
                 value: None,
             },
-            Err(error) => KvResponse::Error {
-                error: Self::map_midge_error(&error),
-            },
+            Err(error) => KvResponse::Error { error },
         }
     }
 
@@ -43,17 +41,18 @@ impl KvActor {
             Ok(tx) => tx,
             Err(error) => return error,
         };
+        if let Err(response) = Self::ensure_writable(active) {
+            return response;
+        }
 
         let scoped_key = Self::encode_scoped_key(&active.scoped_prefix, key);
-        match active.tx.put(scoped_key, value.to_vec(), None) {
+        match active.tx.put(scoped_key, value.to_vec()) {
             Ok(()) => {
                 active.mutation_count = active.mutation_count.saturating_add(1);
                 active.inventory_delta.mark_incomplete();
                 KvResponse::PutOk
             }
-            Err(error) => KvResponse::Error {
-                error: Self::map_midge_error(&error),
-            },
+            Err(error) => KvResponse::Error { error },
         }
     }
 
@@ -68,13 +67,16 @@ impl KvActor {
             Ok(tx) => tx,
             Err(error) => return error,
         };
+        if let Err(response) = Self::ensure_writable(active) {
+            return response;
+        }
 
         let scoped_key = Self::encode_scoped_key(&active.scoped_prefix, key);
         match active.tx.get(&scoped_key) {
             Ok(Some(_)) => KvResponse::Error {
                 error: KvError::AlreadyExists,
             },
-            Ok(None) => match active.tx.put(scoped_key, value.to_vec(), None) {
+            Ok(None) => match active.tx.put(scoped_key, value.to_vec()) {
                 Ok(()) => {
                     active.mutation_count = active.mutation_count.saturating_add(1);
                     active
@@ -82,13 +84,9 @@ impl KvActor {
                         .record_insert(key, key.len() + value.len());
                     KvResponse::InsertOk
                 }
-                Err(error) => KvResponse::Error {
-                    error: Self::map_midge_error(&error),
-                },
+                Err(error) => KvResponse::Error { error },
             },
-            Err(error) => KvResponse::Error {
-                error: Self::map_midge_error(&error),
-            },
+            Err(error) => KvResponse::Error { error },
         }
     }
 
@@ -102,6 +100,9 @@ impl KvActor {
             Ok(tx) => tx,
             Err(error) => return error,
         };
+        if let Err(response) = Self::ensure_writable(active) {
+            return response;
+        }
 
         let scoped_key = Self::encode_scoped_key(&active.scoped_prefix, key);
         match active.tx.delete(scoped_key) {
@@ -110,9 +111,7 @@ impl KvActor {
                 active.inventory_delta.mark_incomplete();
                 KvResponse::DeleteOk
             }
-            Err(error) => KvResponse::Error {
-                error: Self::map_midge_error(&error),
-            },
+            Err(error) => KvResponse::Error { error },
         }
     }
 
@@ -127,6 +126,9 @@ impl KvActor {
             Ok(tx) => tx,
             Err(error) => return error,
         };
+        if let Err(response) = Self::ensure_writable(active) {
+            return response;
+        }
         if start >= end {
             return KvResponse::Error {
                 error: KvError::InvalidRequest("start must be less than end".to_string()),
@@ -141,9 +143,7 @@ impl KvActor {
                 active.inventory_delta.mark_incomplete();
                 KvResponse::DeleteRangeOk
             }
-            Err(error) => KvResponse::Error {
-                error: Self::map_midge_error(&error),
-            },
+            Err(error) => KvResponse::Error { error },
         }
     }
 }
